requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
Pillow
transformers
python-multipart
typing